@@ -92,5 +92,7 @@ DEFAULT_PATH_TRAVERSAL_FUNCTIONS = {
 DBAPI_SQLITE = "sqlite"
 DBAPI_PSYCOPG = "psycopg"
 DBAPI_MYSQL = "mysql"
+DBAPI_MYSQLDB = "mysqldb"
+DBAPI_PYMYSQL = "pymysql"
 DBAPI_MARIADB = "mariadb"
-DBAPI_INTEGRATIONS = (DBAPI_SQLITE, DBAPI_PSYCOPG, DBAPI_MYSQL, DBAPI_MARIADB)
+DBAPI_INTEGRATIONS = (DBAPI_SQLITE, DBAPI_PSYCOPG, DBAPI_MYSQL, DBAPI_MYSQLDB, DBAPI_MARIADB, DBAPI_PYMYSQL)
