@@ -452,9 +452,8 @@ def pytest_load_initial_conftests(early_config, parser, args):
 
 def pytest_configure(config):
     deprecate(
-        "this version of the pytest ddtrace plugin is slated for deprecation",
-        message="set DD_PYTEST_USE_NEW_PLUGIN_BETA=true in your environment to preview the next version of the plugin.",
-        removal_version="3.0.0",
+        "this version of the pytest ddtrace plugin is deprecated",
+        message="remove _DD_PYTEST_USE_LEGACY_PLUGIN from your environment to use the currently supported version.",
         category=DDTraceDeprecationWarning,
     )
     unpatch_unittest()
