@@ -1,7 +1,6 @@
 import os
 import sqlite3
 import sqlite3.dbapi2
-import sys
 
 import wrapt
 
@@ -97,12 +96,9 @@ class TracedSQLite(TracedConnection):
         # sqlite has a few extra sugar functions
         return self.cursor().execute(*args, **kwargs)
 
-    # backup was added in Python 3.7
-    if sys.version_info >= (3, 7, 0):
-
-        def backup(self, target, *args, **kwargs):
-            # sqlite3 checks the type of `target`, it cannot be a wrapped connection
-            # https://github.com/python/cpython/blob/4652093e1b816b78e9a585d671a807ce66427417/Modules/_sqlite/connection.c#L1897-L1899
-            if isinstance(target, TracedConnection):
-                target = target.__wrapped__
-            return self.__wrapped__.backup(target, *args, **kwargs)
+    def backup(self, target, *args, **kwargs):
+        # sqlite3 checks the type of `target`, it cannot be a wrapped connection
+        # https://github.com/python/cpython/blob/4652093e1b816b78e9a585d671a807ce66427417/Modules/_sqlite/connection.c#L1897-L1899
+        if isinstance(target, TracedConnection):
+            target = target.__wrapped__
+        return self.__wrapped__.backup(target, *args, **kwargs)
