@@ -10,8 +10,6 @@ elif sys.version_info >= (3, 11):
     from ddtrace.internal.coverage.instrumentation_py3_11 import instrument_all_lines  # noqa
 elif sys.version_info >= (3, 10):
     from ddtrace.internal.coverage.instrumentation_py3_10 import instrument_all_lines  # noqa
-elif sys.version_info >= (3, 8):
+else:
     # Python 3.8 and 3.9 use the same instrumentation
     from ddtrace.internal.coverage.instrumentation_py3_8 import instrument_all_lines  # noqa
-else:
-    from ddtrace.internal.coverage.instrumentation_py3_7 import instrument_all_lines  # noqa
