@@ -3,7 +3,7 @@ This is the entry point for the IAST instrumentation. `enable_iast_propagation` 
 too but patch_all depends of DD_TRACE_ENABLED environment variable. This is the reason why we need to call it
 here and it's not a duplicate call due to `enable_iast_propagation` has a global variable to avoid multiple calls.
 """
-from ddtrace.appsec._iast._utils import _is_iast_enabled
+from ddtrace.settings.asm import config as asm_config
 
 
 def post_preload():
@@ -11,7 +11,7 @@ def post_preload():
 
 
 def start():
-    if _is_iast_enabled():
+    if asm_config._iast_enabled:
         from ddtrace.appsec._iast import enable_iast_propagation
 
         enable_iast_propagation()
