@@ -1,15 +1,11 @@
-def get_version():
-    # type: () -> str
+def get_version() -> str:
     try:
         from ._version import version
 
         return version
     except ImportError:
-        try:
-            from importlib.metadata import version as ilm_version
-        except ImportError:
-            # required for python3.7
-            from importlib_metadata import version as ilm_version  # type: ignore[no-redef]
+        from importlib.metadata import version as ilm_version
+
         try:
             return ilm_version("ddtrace")
         except ModuleNotFoundError:
