@@ -11,7 +11,6 @@ latest = ""
 
 
 SUPPORTED_PYTHON_VERSIONS: List[Tuple[int, int]] = [
-    (3, 7),
     (3, 8),
     (3, 9),
     (3, 10),
@@ -24,8 +23,6 @@ SUPPORTED_PYTHON_VERSIONS: List[Tuple[int, int]] = [
 def version_to_str(version: Tuple[int, int]) -> str:
     """Convert a Python version tuple to a string
 
-    >>> version_to_str((3, 7))
-    '3.7'
     >>> version_to_str((3, 8))
     '3.8'
     >>> version_to_str((3, 9))
@@ -45,8 +42,6 @@ def version_to_str(version: Tuple[int, int]) -> str:
 def str_to_version(version: str) -> Tuple[int, int]:
     """Convert a Python version string to a tuple
 
-    >>> str_to_version("3.7")
-    (3, 7)
     >>> str_to_version("3.8")
     (3, 8)
     >>> str_to_version("3.9")
@@ -71,13 +66,13 @@ def select_pys(min_version: str = MIN_PYTHON_VERSION, max_version: str = MAX_PYT
     """Helper to select python versions from the list of versions we support
 
     >>> select_pys()
-    ['3.7', '3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
+    ['3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
     >>> select_pys(min_version='3')
-    ['3.7', '3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
+    ['3.8', '3.9', '3.10', '3.11', '3.12', '3.13']
     >>> select_pys(max_version='3')
     []
-    >>> select_pys(min_version='3.7', max_version='3.9')
-    ['3.7', '3.8', '3.9']
+    >>> select_pys(min_version='3.8', max_version='3.9')
+    ['3.8', '3.9']
     """
     min_version = str_to_version(min_version)
     max_version = str_to_version(max_version)
@@ -103,7 +98,6 @@ venv = Venv(
         "DD_INJECT_FORCE": "1",
         "DD_PATCH_MODULES": "unittest:false",
         "CMAKE_BUILD_PARALLEL_LEVEL": "12",
-        "DD_PYTEST_USE_NEW_PLUGIN_BETA": "true",
     },
     venvs=[
         Venv(
@@ -154,6 +148,8 @@ venv = Venv(
                 "simplejson": latest,
                 "SQLAlchemy": "==2.0.22",
                 "psycopg2-binary": "~=2.9.9",
+                "pymysql": latest,
+                "mysqlclient": "==2.1.1",
                 "googleapis-common-protos": latest,
                 "grpcio": latest,
             },
@@ -386,12 +382,6 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys="3.7",
-                    pkgs={
-                        "pytest-asyncio": "~=0.21.1",
-                    },
-                ),
-                Venv(
                     pys=select_pys(min_version="3.8", max_version="3.11"),
                     pkgs={
                         "pytest-asyncio": "~=0.23.7",
@@ -425,7 +415,7 @@ venv = Venv(
                     },
                     venvs=[
                         Venv(
-                            pys=select_pys(min_version="3.7", max_version="3.8"),
+                            pys="3.8",
                             pkgs={
                                 "gevent": "~=20.12.0",
                                 # greenlet v1.0.0 adds support for contextvars
@@ -522,7 +512,7 @@ venv = Venv(
             # FIXME: tests fail on vertica 1.x
             # Venv(
             #     # vertica-python added support for Python 3.9/3.10 in 1.0
-            #     pys=select_pys(min_version="3.7", max_version="3.10"),
+            #     pys=select_pys(min_version="3.8", max_version="3.10"),
             #     pkgs={"vertica-python": ["~=1.0", latest]},
             # ),
             # Venv(
@@ -573,7 +563,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.12"),
+                    pys=select_pys(min_version="3.8", max_version="3.12"),
                     pkgs={
                         "falcon": [
                             "~=3.0.0",
@@ -631,25 +621,6 @@ venv = Venv(
             venvs=[
                 # Celery 4.3 wants Kombu >= 4.4 and Redis >= 3.2
                 # Split into <3.8 and >=3.8 to pin importlib_metadata dependency for kombu
-                Venv(
-                    # celery dropped support for Python 2.7/3.5 in 5.0
-                    pkgs={
-                        "pytest": "~=4.0",
-                        "celery": [
-                            latest,  # most recent 4.x
-                        ],
-                        "redis": "~=3.5",
-                        "kombu": "~=4.4",
-                        "importlib_metadata": "<5.0",  # kombu using deprecated shims removed in importlib_metadata 5.0
-                        "pytest-cov": "~=3.0",
-                        "pytest-mock": "==2.0.0",
-                    },
-                    venvs=[
-                        # exceptiongroup latest specified to avoid riot bug: https://github.com/DataDog/riot/issues/211
-                        Venv(pys="3.7", pkgs={"exceptiongroup": latest}),
-                    ],
-                ),
-                # Venv(
                 #     # celery added support for Python 3.9 in 4.x
                 #     pys=select_pys(min_version="3.8", max_version="3.9"),
                 #     pkgs={
@@ -663,21 +634,6 @@ venv = Venv(
                 # ),
                 # Celery 5.x wants Python 3.6+
                 # Split into <3.8 and >=3.8 to pin importlib_metadata dependency for kombu
-                Venv(
-                    pys="3.7",
-                    env={
-                        # https://docs.celeryproject.org/en/v5.0.5/userguide/testing.html#enabling
-                        "PYTEST_PLUGINS": "celery.contrib.pytest",
-                    },
-                    pkgs={
-                        "celery": [
-                            "~=5.1.0",
-                            latest,
-                        ],
-                        "redis": "~=3.5",
-                        "importlib_metadata": "<5.0",  # kombu using deprecated shims removed in importlib_metadata 5.0
-                    },
-                ),
                 Venv(
                     pys=select_pys(min_version="3.8", max_version="3.9"),
                     env={
@@ -725,7 +681,7 @@ venv = Venv(
                 ),
                 Venv(
                     # cherrypy added support for Python 3.11 in 18.7
-                    pys=select_pys(min_version="3.7"),
+                    pys=select_pys(min_version="3.8"),
                     pkgs={
                         "cherrypy": [">=18.0,<19", latest],
                         "more_itertools": "<8.11.0",
@@ -745,7 +701,7 @@ venv = Venv(
                 # pymmongo<3.9, 3.9<=pymongo<3.12, 3.12<=pymongo<4.5, pymongo>=4.5
                 # To get full test coverage we must test all these version ranges
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={"pymongo": ["~=3.8.0", "~=3.9.0", "~=3.11", "~=4.0", latest]},
                 ),
                 Venv(
@@ -789,14 +745,6 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    # django dropped support for Python 3.6/3.7 in 4.0
-                    pys=select_pys(max_version="3.7"),
-                    pkgs={
-                        "django": "~=3.2",
-                        "channels": ["~=3.0", latest],
-                    },
-                ),
-                Venv(
                     # django dropped support for Python 3.8/3.9 in 5.0
                     pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
@@ -827,7 +775,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7"),
+                    pys=select_pys(min_version="3.8"),
                     pkgs={
                         "django_hosts": "~=4.0",
                         "django": "~=3.2",
@@ -852,14 +800,14 @@ venv = Venv(
             venvs=[
                 Venv(
                     # djangorestframework dropped support for Django 2.x in 3.14
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "django": ">=2.2,<2.3",
                         "djangorestframework": ["==3.12.4", "==3.13.1"],
                     },
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7"),
+                    pys=select_pys(min_version="3.8"),
                     pkgs={
                         "django": "~=3.2",
                         "djangorestframework": ">=3.11,<3.12",
@@ -868,7 +816,7 @@ venv = Venv(
                 Venv(
                     pys=select_pys(min_version="3.8"),
                     pkgs={
-                        "django": "~=4.0",
+                        "django": ["~=4.0"],
                         "djangorestframework": ["~=3.13", latest],
                     },
                 ),
@@ -1028,26 +976,11 @@ venv = Venv(
                         "werkzeug": "<2.0",
                     },
                 ),
-                # Flask 2.x.x
                 Venv(
-                    pys=select_pys(min_version="3.7"),
-                    pkgs={"flask": "~=2.0"},
-                ),
-                Venv(
-                    pys=select_pys(max_version="3.7"),
-                    command="python tests/ddtrace_run.py pytest {cmdargs} tests/contrib/flask_autopatch",
-                    env={
-                        "DD_SERVICE": "test.flask.service",
-                        "DD_PATCH_MODULES": "jinja2:false",
-                    },
-                    pkgs={"flask": "~=2.0"},
-                ),
-                # Flask 3.x.x
-                Venv(
-                    # Flask 3.x.x dropped support for Python 3.7
                     pys=select_pys(min_version="3.8"),
                     pkgs={
                         "flask": [
+                            "~=2.0",
                             "~=3.0.0",
                             latest,
                         ],
@@ -1102,7 +1035,6 @@ venv = Venv(
                         "markupsafe": "<2.0",
                     },
                     venvs=[
-                        Venv(pys=select_pys(max_version="3.7")),
                         Venv(pys=select_pys(min_version="3.8", max_version="3.9"), pkgs={"exceptiongroup": latest}),
                     ],
                 ),
@@ -1119,7 +1051,7 @@ venv = Venv(
                     },
                     venvs=[
                         Venv(
-                            pys=select_pys(min_version="3.7", max_version="3.11"),
+                            pys=select_pys(min_version="3.8", max_version="3.11"),
                         ),
                         Venv(pys=select_pys(min_version="3.12"), pkgs={"redis": latest}),
                     ],
@@ -1131,7 +1063,7 @@ venv = Venv(
                     },
                     venvs=[
                         Venv(
-                            pys=select_pys(min_version="3.7", max_version="3.11"),
+                            pys=select_pys(min_version="3.8", max_version="3.11"),
                         ),
                         Venv(pys=select_pys(min_version="3.12"), pkgs={"redis": latest}),
                     ],
@@ -1155,7 +1087,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={"mysql-connector-python": ["==8.0.5", latest]},
                 ),
                 Venv(
@@ -1182,11 +1114,11 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.8"),
+                    pys="3.8",
                     pkgs={"psycopg2-binary": "~=2.8.0"},
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.12"),
+                    pys=select_pys(min_version="3.8", max_version="3.12"),
                     # psycopg2-binary added support for Python 3.9/3.10 in 2.9.1
                     # psycopg2-binary added support for Python 3.11 in 2.9.2
                     pkgs={"psycopg2-binary": ["~=2.9.2", latest]},
@@ -1204,7 +1136,7 @@ venv = Venv(
                     pkgs={"psycopg": [latest]},
                     venvs=[
                         Venv(
-                            pys=select_pys(min_version="3.7", max_version="3.11"),
+                            pys=select_pys(min_version="3.8", max_version="3.11"),
                             pkgs={
                                 "pytest-asyncio": "==0.21.1",
                             },
@@ -1241,7 +1173,7 @@ venv = Venv(
             # TODO: Py312 requires changes to test code
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.11"),
+                    pys=select_pys(min_version="3.8", max_version="3.11"),
                     pkgs={
                         "pynamodb": ["~=5.0", "~=5.3", "<6.0"],
                         "moto": ">=1.0,<2.0",
@@ -1270,10 +1202,6 @@ venv = Venv(
             venvs=[
                 # starlette added new TestClient after v0.20
                 # starlette added new root_path/path definitions after v0.33
-                Venv(
-                    pys="3.7",
-                    pkgs={"starlette": ["~=0.14.0", "~=0.20.0", latest]},
-                ),
                 Venv(
                     # starlette added support for Python 3.9 in 0.14
                     pys=select_pys(min_version="3.8", max_version="3.9"),
@@ -1315,7 +1243,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.12"),
+                    pys=select_pys(min_version="3.8", max_version="3.12"),
                     pkgs={
                         "greenlet": "==3.0.3",
                         "sqlalchemy": ["~=1.3.0", latest],
@@ -1338,16 +1266,6 @@ venv = Venv(
                 "requests-mock": ">=1.4",
             },
             venvs=[
-                # requests added support for Python 3.7 in 2.20
-                Venv(
-                    pys="3.7",
-                    pkgs={
-                        "requests": [
-                            "~=2.20.0",
-                            latest,
-                        ],
-                    },
-                ),
                 Venv(
                     # requests added support for Python 3.8 in 2.23
                     pys="3.8",
@@ -1436,7 +1354,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.8"),
+                    pys="3.8",
                     pkgs={"mongoengine": ["~=0.23", latest]},
                 ),
                 Venv(
@@ -1454,7 +1372,7 @@ venv = Venv(
                 "asgiref": ["~=3.0.0", "~=3.0", latest],
                 "pytest-randomly": latest,
             },
-            pys=select_pys(min_version="3.7"),
+            pys=select_pys(min_version="3.8"),
             command="pytest {cmdargs} tests/contrib/asgi",
         ),
         Venv(
@@ -1465,7 +1383,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.10"),
+                    pys=select_pys(min_version="3.8", max_version="3.10"),
                     pkgs={
                         "mariadb": [
                             "~=1.0.0",
@@ -1490,7 +1408,7 @@ venv = Venv(
                     pkgs={"pymysql": "~=0.10"},
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.12"),
+                    pys=select_pys(min_version="3.8", max_version="3.12"),
                     pkgs={
                         "pymysql": [
                             "~=1.0",
@@ -1511,7 +1429,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "pyramid": [
                             "~=1.10",
@@ -1540,7 +1458,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.11"),
+                    pys=select_pys(min_version="3.8", max_version="3.11"),
                     pkgs={
                         "aiobotocore": ["~=1.4.2", "~=2.0.0", latest],
                     },
@@ -1564,7 +1482,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.10"),
+                    pys=select_pys(min_version="3.8", max_version="3.10"),
                     pkgs={"fastapi": ["~=0.64.0", "~=0.90.0", latest]},
                 ),
                 Venv(
@@ -1579,7 +1497,7 @@ venv = Venv(
             command="pytest {cmdargs} tests/contrib/aiomysql",
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.12"),
+                    pys=select_pys(min_version="3.8", max_version="3.12"),
                     pkgs={
                         "pytest-randomly": latest,
                         "pytest-asyncio": "==0.21.1",
@@ -1604,11 +1522,10 @@ venv = Venv(
             },
             env={
                 "DD_AGENT_PORT": "9126",
-                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "pytest": [
                             ">=6.0,<7.0",
@@ -1634,12 +1551,12 @@ venv = Venv(
                             venvs=[
                                 Venv(
                                     env={
-                                        "DD_PYTEST_USE_NEW_PLUGIN_BETA": "0",
+                                        "_DD_PYTEST_USE_LEGACY_PLUGIN": "true",
                                     },
                                 ),
                                 Venv(
                                     env={
-                                        "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
+                                        "_DD_PYTEST_USE_LEGACY_PLUGIN": "false",
                                     },
                                 ),
                             ],
@@ -1662,12 +1579,12 @@ venv = Venv(
                     venvs=[
                         Venv(
                             env={
-                                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "0",
+                                "DD_PYTEST_LEGACY_PLUGIN": "true",
                             },
                         ),
                         Venv(
                             env={
-                                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
+                                "_DD_PYTEST_USE_LEGACY_PLUGIN": "false",
                             },
                         ),
                     ],
@@ -1698,7 +1615,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "pytest": [
                             ">=6.0,<7.0",
@@ -1723,7 +1640,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "pytest-bdd": [
                             ">=4.0,<5.0",
@@ -1734,12 +1651,12 @@ venv = Venv(
                     venvs=[
                         Venv(
                             env={
-                                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "0",
+                                "_DD_PYTEST_USE_LEGACY_PLUGIN": "true",
                             },
                         ),
                         Venv(
                             env={
-                                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
+                                "_DD_PYTEST_USE_LEGACY_PLUGIN": "false",
                             },
                         ),
                     ],
@@ -1755,12 +1672,12 @@ venv = Venv(
                     venvs=[
                         Venv(
                             env={
-                                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "0",
+                                "_DD_PYTEST_USE_LEGACY_PLUGIN": "true",
                             },
                         ),
                         Venv(
                             env={
-                                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
+                                "_DD_PYTEST_USE_LEGACY_PLUGIN": "false",
                             },
                         ),
                     ],
@@ -1769,7 +1686,7 @@ venv = Venv(
         ),
         Venv(
             name="pytest-benchmark",
-            pys=select_pys(min_version="3.7", max_version="3.12"),
+            pys=select_pys(min_version="3.8", max_version="3.12"),
             command="pytest {cmdargs} --no-ddtrace --no-cov tests/contrib/pytest_benchmark/",
             pkgs={
                 "msgpack": latest,
@@ -1783,7 +1700,7 @@ venv = Venv(
                         ]
                     },
                     env={
-                        "DD_PYTEST_USE_NEW_PLUGIN_BETA": "0",
+                        "_DD_PYTEST_USE_LEGACY_PLUGIN": "true",
                     },
                 ),
                 Venv(
@@ -1793,7 +1710,7 @@ venv = Venv(
                         ]
                     },
                     env={
-                        "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
+                        "_DD_PYTEST_USE_LEGACY_PLUGIN": "false",
                     },
                 ),
             ],
@@ -1809,7 +1726,7 @@ venv = Venv(
                 # Versions between 1.14 and 1.20 have known threading issues
                 # See https://github.com/grpc/grpc/issues/18994
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={"grpcio": ["~=1.34.0", latest]},
                 ),
                 Venv(
@@ -1851,13 +1768,6 @@ venv = Venv(
             env={"_DD_TRACE_GRPC_AIO_ENABLED": "true"},
             venvs=[
                 Venv(
-                    pys="3.7",
-                    pkgs={
-                        "grpcio": ["~=1.34.0", "~=1.59.0"],
-                        "pytest-asyncio": "==0.21.1",
-                    },
-                ),
-                Venv(
                     pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "grpcio": ["~=1.34.0", "~=1.59.0"],
@@ -1886,7 +1796,7 @@ venv = Venv(
         Venv(
             name="graphene",
             command="pytest {cmdargs} tests/contrib/graphene",
-            pys=select_pys(min_version="3.7"),
+            pys=select_pys(min_version="3.8"),
             pkgs={
                 "graphene": ["~=3.0.0", latest],
                 "pytest-asyncio": "==0.21.1",
@@ -1897,7 +1807,7 @@ venv = Venv(
         Venv(
             name="graphql",
             command="pytest {cmdargs} tests/contrib/graphql",
-            pys=select_pys(min_version="3.7"),
+            pys=select_pys(min_version="3.8"),
             pkgs={
                 "pytest-asyncio": "==0.21.1",
                 "graphql-core": ["~=3.2.0", latest],
@@ -1913,7 +1823,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.8"),
+                    pys="3.8",
                     pkgs={
                         "rq": [
                             "~=1.8.0",
@@ -1946,7 +1856,7 @@ venv = Venv(
         ),
         Venv(
             name="httpx",
-            pys=select_pys(min_version="3.7"),
+            pys=select_pys(min_version="3.8"),
             command="pytest {cmdargs} tests/contrib/httpx",
             pkgs={
                 "pytest-asyncio": "==0.21.1",
@@ -1965,12 +1875,6 @@ venv = Venv(
                 "pytest-randomly": latest,
             },
             venvs=[
-                Venv(
-                    # Support added for Python 3.7 in 1.25.0
-                    # Support removed for Python 3.7 after 1.26.0
-                    pys="3.7",
-                    pkgs={"urllib3": ["==1.25.0", "~=1.26.0"]},
-                ),
                 Venv(
                     # Support added for Python 3.8 in 1.25.0
                     pys="3.8",
@@ -1999,11 +1903,8 @@ venv = Venv(
             ],
         ),
         Venv(
-            # cassandra-driver does not officially support 3.9, 3.10
-            # releases 3.7 and 3.8 are broken on Python >= 3.7
-            # (see https://github.com/r4fek/django-cassandra-engine/issues/104)
             name="cassandra",
-            pys=select_pys(max_version="3.8"),
+            pys="3.8",  # see https://github.com/r4fek/django-cassandra-engine/issues/104
             pkgs={"cassandra-driver": ["~=3.24.0", latest], "pytest-randomly": latest},
             command="pytest {cmdargs} tests/contrib/cassandra",
         ),
@@ -2013,7 +1914,7 @@ venv = Venv(
             pkgs={"urllib3": "~=1.26.15", "pytest-randomly": latest},
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.8"),
+                    pys="3.8",
                     pkgs={"algoliasearch": ["~=2.5", "~=2.6"]},
                 ),
                 Venv(
@@ -2026,7 +1927,7 @@ venv = Venv(
         Venv(
             name="aiopg",
             command="pytest {cmdargs} tests/contrib/aiopg",
-            pys=select_pys(min_version="3.7", max_version="3.9"),
+            pys=select_pys(min_version="3.8", max_version="3.9"),
             pkgs={
                 "sqlalchemy": latest,
                 "aiopg": "~=0.16.0",
@@ -2034,7 +1935,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7"),
+                    pys=select_pys(min_version="3.8"),
                     pkgs={
                         "aiopg": ["~=1.0", "~=1.4.0"],
                     },
@@ -2054,12 +1955,6 @@ venv = Venv(
                 "yarl": "~=1.0",
             },
             venvs=[
-                Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.7"),
-                    pkgs={
-                        "pytest-asyncio": ["==0.21.1"],
-                    },
-                ),
                 Venv(
                     pys=select_pys(min_version="3.8"),
                     pkgs={
@@ -2086,12 +1981,6 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.7"),
-                    pkgs={
-                        "pytest-asyncio": ["==0.21.1"],
-                    },
-                ),
-                Venv(
                     pys=select_pys(min_version="3.8"),
                     pkgs={
                         "pytest-asyncio": ["==0.23.7"],
@@ -2115,7 +2004,7 @@ venv = Venv(
                     },
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7"),
+                    pys=select_pys(min_version="3.8"),
                     pkgs={
                         "jinja2": ["~=3.0.0", latest],
                     },
@@ -2148,12 +2037,6 @@ venv = Venv(
                     },
                     venvs=[
                         Venv(
-                            pys="3.7",
-                            pkgs={
-                                "pytest-asyncio": "==0.21.1",
-                            },
-                        ),
-                        Venv(
                             pys=select_pys(min_version="3.8", max_version="3.10"),
                             pkgs={
                                 "pytest-asyncio": "==0.23.7",
@@ -2182,7 +2065,7 @@ venv = Venv(
         ),
         Venv(
             name="aredis",
-            pys=select_pys(min_version="3.7", max_version="3.9"),
+            pys=select_pys(min_version="3.8", max_version="3.9"),
             command="pytest {cmdargs} tests/contrib/aredis",
             pkgs={
                 "pytest-asyncio": "==0.21.1",
@@ -2192,7 +2075,7 @@ venv = Venv(
         ),
         Venv(
             name="avro",
-            pys=select_pys(min_version="3.7"),
+            pys=select_pys(min_version="3.8"),
             command="pytest {cmdargs} tests/contrib/avro",
             pkgs={
                 "avro": latest,
@@ -2217,7 +2100,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={"yaaredis": ["~=2.0.0", latest]},
                 ),
                 Venv(
@@ -2239,14 +2122,14 @@ venv = Venv(
             venvs=[
                 Venv(
                     # sanic added support for Python 3.9 in 20.12
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "sanic": "~=20.12",
                         "pytest-sanic": "~=1.6.2",
                     },
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
                         "sanic": [
                             "~=21.3",
@@ -2264,7 +2147,7 @@ venv = Venv(
                     },
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.10"),
+                    pys=select_pys(min_version="3.8", max_version="3.10"),
                     pkgs={
                         "sanic": ["~=22.3", "~=22.12"],
                         "sanic-testing": "~=22.3.0",
@@ -2297,7 +2180,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.8"),
+                    pys="3.8",
                     pkgs={"snowflake-connector-python": ["~=2.3.0", "~=2.9.0", latest]},
                 ),
                 Venv(
@@ -2335,7 +2218,7 @@ venv = Venv(
             venvs=[
                 # our test_asyncpg.py uses `yield` in an async function and is not compatible with Python 3.5
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.8"),
+                    pys="3.8",
                     pkgs={"asyncpg": ["~=0.23", latest]},
                 ),
                 Venv(
@@ -2358,15 +2241,6 @@ venv = Venv(
                     pkgs={"asyncpg": [latest]},
                 ),
             ],
-        ),
-        Venv(
-            name="asyncio",
-            command="pytest {cmdargs} tests/contrib/asyncio",
-            pys=select_pys(),
-            pkgs={
-                "pytest-randomly": latest,
-                "pytest-asyncio": "==0.21.1",
-            },
         ),
         Venv(
             name="futures",
@@ -2415,7 +2289,6 @@ venv = Venv(
                 "pytest-randomly": latest,
             },
             venvs=[
-                Venv(pys=["3.7"]),
                 Venv(pys=select_pys(min_version="3.8", max_version="3.10")),
                 Venv(pys=select_pys(min_version="3.11"), pkgs={"attrs": latest}),
             ],
@@ -2428,7 +2301,7 @@ venv = Venv(
             },
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.10"),
+                    pys=select_pys(min_version="3.8", max_version="3.10"),
                     pkgs={
                         "dogpile.cache": [
                             "~=0.9",
@@ -2483,6 +2356,15 @@ venv = Venv(
             },
         ),
         Venv(
+            name="asyncio",
+            command="pytest {cmdargs} tests/contrib/asyncio",
+            pys=select_pys(),
+            pkgs={
+                "pytest-randomly": latest,
+                "pytest-asyncio": "==0.21.1",
+            },
+        ),
+        Venv(
             name="openai",
             command="pytest {cmdargs} tests/contrib/openai",
             pkgs={
@@ -2520,12 +2402,12 @@ venv = Venv(
                     command="pytest {cmdargs} tests/opentracer/core",
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7"),
+                    pys=select_pys(min_version="3.8"),
                     command="pytest {cmdargs} tests/opentracer/test_tracer_asyncio.py",
                     pkgs={"pytest-asyncio": "==0.21.1"},
                 ),
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.11"),
+                    pys=select_pys(min_version="3.8", max_version="3.11"),
                     command="pytest {cmdargs} tests/opentracer/test_tracer_tornado.py",
                     # TODO: update opentracing tests to be compatible with Tornado v6.
                     # https://github.com/opentracing/opentracing-python/issues/136
@@ -2537,7 +2419,7 @@ venv = Venv(
                     command="pytest {cmdargs} tests/opentracer/test_tracer_gevent.py",
                     venvs=[
                         Venv(
-                            pys=select_pys(min_version="3.7", max_version="3.8"),
+                            pys="3.8",
                             pkgs={
                                 "gevent": latest,
                                 "greenlet": latest,
@@ -2591,7 +2473,7 @@ venv = Venv(
             venvs=[
                 Venv(
                     # pylibmc added support for Python 3.8/3.9/3.10 in 1.6.2
-                    pys=select_pys(min_version="3.7", max_version="3.10"),
+                    pys=select_pys(min_version="3.8", max_version="3.10"),
                     pkgs={
                         "pylibmc": ["~=1.6.2", latest],
                     },
@@ -2609,15 +2491,6 @@ venv = Venv(
             command="pytest {cmdargs} tests/contrib/kombu",
             pkgs={"pytest-randomly": latest},
             venvs=[
-                # Kombu>=4.2 only supports Python 3.7+
-                Venv(
-                    pys="3.7",
-                    pkgs={
-                        "kombu": [">=4.6,<4.7", ">=5.0,<5.1", latest],
-                        # kombu using deprecated shims removed in importlib-metadata 5.0 pre-Python 3.8
-                        "importlib_metadata": "<5.0",
-                    },
-                ),
                 Venv(
                     pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={
@@ -2641,7 +2514,7 @@ venv = Venv(
             venvs=[
                 Venv(
                     # tornado added support for Python 3.8/3.9 in 6.1
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={"tornado": ["~=6.1", "~=6.2"]},
                 ),
                 Venv(
@@ -2662,7 +2535,7 @@ venv = Venv(
             pkgs={"pytest-randomly": latest},
             venvs=[
                 Venv(
-                    pys=select_pys(min_version="3.7", max_version="3.9"),
+                    pys=select_pys(min_version="3.8", max_version="3.9"),
                     pkgs={"mysqlclient": ["~=2.0", "~=2.1", latest]},
                 ),
                 Venv(
@@ -2845,7 +2718,7 @@ venv = Venv(
                     command="pytest {cmdargs} -vv tests/contrib/kafka",
                     venvs=[
                         Venv(
-                            pys=select_pys(min_version="3.7", max_version="3.10"),
+                            pys=select_pys(min_version="3.8", max_version="3.10"),
                             pkgs={"confluent-kafka": ["~=1.9.2", latest]},
                         ),
                         # confluent-kafka added support for Python 3.11 in 2.0.2
@@ -2869,7 +2742,7 @@ venv = Venv(
         Venv(
             name="azure_functions",
             command="pytest {cmdargs} tests/contrib/azure_functions",
-            pys=select_pys(min_version="3.7", max_version="3.11"),
+            pys=select_pys(min_version="3.8", max_version="3.11"),
             pkgs={
                 "azure.functions": latest,
                 "requests": latest,
@@ -2949,28 +2822,6 @@ venv = Venv(
                 "pytest-randomly": latest,
             },
             venvs=[
-                # Python 3.7
-                Venv(
-                    pys="3.7",
-                    pkgs={"uwsgi": latest},
-                    venvs=[
-                        Venv(
-                            pkgs={
-                                "protobuf": ["==3.8.0", latest],
-                            },
-                        ),
-                        # Gevent
-                        Venv(
-                            env={
-                                "DD_PROFILE_TEST_GEVENT": "1",
-                            },
-                            pkgs={
-                                "gunicorn[gevent]": latest,
-                                "gevent": latest,
-                            },
-                        ),
-                    ],
-                ),
                 # Python 3.8 + 3.9
                 Venv(
                     pys=["3.8", "3.9"],
@@ -3072,8 +2923,6 @@ venv = Venv(
             env={
                 "DD_PROFILING_ENABLE_ASSERTS": "1",
                 "DD_PROFILING_EXPORT_LIBDD_ENABLED": "1",
-                # Enable pytest v2 plugin to handle pytest-cpp items in the test suite
-                "DD_PYTEST_USE_NEW_PLUGIN_BETA": "1",
                 "CPUCOUNT": "12",
             },
             pkgs={
@@ -3089,28 +2938,6 @@ venv = Venv(
                 "pytest-randomly": latest,
             },
             venvs=[
-                # Python 3.7
-                Venv(
-                    pys="3.7",
-                    pkgs={"uwsgi": latest},
-                    venvs=[
-                        Venv(
-                            pkgs={
-                                "protobuf": ["==3.8.0", latest],
-                            },
-                        ),
-                        # Gevent
-                        Venv(
-                            env={
-                                "DD_PROFILE_TEST_GEVENT": "1",
-                            },
-                            pkgs={
-                                "gunicorn[gevent]": latest,
-                                "gevent": latest,
-                            },
-                        ),
-                    ],
-                ),
                 # Python 3.8 + 3.9
                 Venv(
                     pys=["3.8", "3.9"],
