@@ -4,8 +4,8 @@ import logging
 import os
 import sys
 
-from ddtrace.appsec._iast._utils import _is_iast_enabled
 from ddtrace.ext import SpanTypes
+from ddtrace.settings.asm import config as asm_config
 from ddtrace.trace import tracer
 
 
@@ -29,11 +29,11 @@ if __name__ == "__main__":
     main()
     if not iast_enabled:
         # Disabled by env var but then enabled with ``tracer.configure``
-        assert _is_iast_enabled()
+        assert asm_config._iast_enabled
         assert "ddtrace.appsec._iast.processor" in sys.modules
     else:
         # Enabled by env var but then disabled with ``tracer.configure``
-        assert not _is_iast_enabled()
+        assert not asm_config._iast_enabled
         # Module was loaded before
         assert "ddtrace.appsec._iast.processor" in sys.modules
         # But processor is not used by the tracer
