@@ -1,5 +1,6 @@
 import pytest
 
+from ddtrace import patch
 from ddtrace.appsec._iast._taint_tracking import OriginType
 from ddtrace.appsec._iast._taint_tracking._taint_objects import is_pyobject_tainted
 from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
@@ -16,6 +17,14 @@ DDBBS = [
         "tests.appsec.iast.fixtures.taint_sinks.sql_injection_sqlite3",
     ),
     (
+        "tests/appsec/iast/fixtures/taint_sinks/sql_injection_mysqldb.py",
+        "tests.appsec.iast.fixtures.taint_sinks.sql_injection_mysqldb",
+    ),
+    (
+        "tests/appsec/iast/fixtures/taint_sinks/sql_injection_pymysql.py",
+        "tests.appsec.iast.fixtures.taint_sinks.sql_injection_pymysql",
+    ),
+    (
         "tests/appsec/iast/fixtures/taint_sinks/sql_injection_psycopg2.py",
         "tests.appsec.iast.fixtures.taint_sinks.sql_injection_psycopg2",
     ),
@@ -24,6 +33,10 @@ DDBBS = [
         "tests.appsec.iast.fixtures.taint_sinks.sql_injection_sqlalchemy",
     ),
 ]
+
+
+def setup_module():
+    patch(pymysql=True, mysqldb=True)
 
 
 @pytest.mark.parametrize("fixture_path,fixture_module", DDBBS)
@@ -39,7 +52,12 @@ def test_sql_injection(fixture_path, fixture_module, iast_context_defaults):
 
     mod.sqli_simple(table)
     data = _get_iast_data()
-    vulnerability = data["vulnerabilities"][0]
+    assert len(data["vulnerabilities"]) >= 1
+    # We will pick up weak hash vulnerabilities in some db connector libraries
+    # but we are only interested in SQL Injection vulnerabilities
+    sqli_vulnerabilities = [x for x in data["vulnerabilities"] if x["type"] == VULN_SQL_INJECTION]
+    assert len(sqli_vulnerabilities) == 1
+    vulnerability = sqli_vulnerabilities[0]
     source = data["sources"][0]
     assert vulnerability["type"] == VULN_SQL_INJECTION
     assert vulnerability["evidence"]["valueParts"] == [
@@ -74,5 +92,8 @@ def test_sql_injection_deduplication(fixture_path, fixture_module, iast_context_
         mod.sqli_simple(table)
 
     data = _get_iast_data()
-    assert len(data["vulnerabilities"]) == 1
+    # We will pick up weak hash vulnerabilities in some db connector libraries
+    # but we are only interested in SQL Injection vulnerabilities
+    sqli_vulnerabilities = [x for x in data["vulnerabilities"] if x["type"] == VULN_SQL_INJECTION]
+    assert len(sqli_vulnerabilities) == 1
     VulnerabilityBase._prepare_report._reset_cache()
