@@ -4,12 +4,7 @@ from ddtrace.trace import tracer
 from tests.webclient import PingFilter
 
 
-tracer.configure(
-    settings={
-        "FILTERS": [PingFilter()],
-    }
-)
-
+tracer._configure(trace_processors=[PingFilter()])
 
 ALLOWED_HOSTS = [
     "testserver",
