@@ -7,7 +7,6 @@ from ddtrace.appsec._asm_request_context import start_context
 from ddtrace.appsec._constants import IAST
 from ddtrace.appsec._iast import oce
 from ddtrace.appsec._iast._patch_modules import patch_iast
-from ddtrace.appsec._iast._utils import _is_python_version_supported as python_supported_by_iast
 from ddtrace.appsec._iast.constants import VULN_CMDI
 from ddtrace.appsec._iast.constants import VULN_HEADER_INJECTION
 from ddtrace.appsec._iast.constants import VULN_INSECURE_COOKIE
@@ -15,6 +14,7 @@ from ddtrace.appsec._iast.constants import VULN_SQL_INJECTION
 from ddtrace.appsec._iast.constants import VULN_STACKTRACE_LEAK
 from ddtrace.ext import SpanTypes
 from ddtrace.internal.compat import urlencode
+from ddtrace.settings.asm import config as asm_config
 from tests.appsec.iast.iast_utils import get_line_and_hash
 from tests.utils import override_env
 from tests.utils import override_global_config
@@ -83,7 +83,7 @@ def _aux_appsec_get_root_span_with_exception(
         return False
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_weak_hash(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -96,7 +96,7 @@ def test_django_weak_hash(client, test_spans, tracer):
         assert vulnerability["evidence"]["value"] == "md5"
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_tainted_user_agent_iast_enabled(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -137,7 +137,7 @@ def test_django_tainted_user_agent_iast_enabled(client, test_spans, tracer):
         50.0,
     ],
 )
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_view_with_exception(client, test_spans, tracer, payload, content_type, deduplication, sampling):
     with override_global_config(
         dict(_iast_enabled=True, _deduplication_enabled=deduplication, _iast_request_sampling=sampling)
@@ -153,7 +153,7 @@ def test_django_view_with_exception(client, test_spans, tracer, payload, content
         assert response is False
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_tainted_user_agent_iast_disabled(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -175,7 +175,7 @@ def test_django_tainted_user_agent_iast_disabled(client, test_spans, tracer):
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_tainted_user_agent_iast_enabled_sqli_http_request_parameter(client, test_spans, tracer):
     with override_global_config(
         dict(_iast_enabled=True, _iast_deduplication_enabled=False, _iast_request_sampling=100.0)
@@ -224,7 +224,7 @@ def test_django_tainted_user_agent_iast_enabled_sqli_http_request_parameter(clie
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_request_parameter_name_get(client, test_spans, tracer):
     with override_global_config(
         dict(_iast_enabled=True, _iast_deduplication_enabled=False, _iast_request_sampling=100.0)
@@ -275,7 +275,7 @@ def test_django_sqli_http_request_parameter_name_get(client, test_spans, tracer)
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_request_parameter_name_post(client, test_spans, tracer):
     with override_global_config(
         dict(_iast_enabled=True, _iast_deduplication_enabled=False, _iast_request_sampling=100.0)
@@ -327,7 +327,7 @@ def test_django_sqli_http_request_parameter_name_post(client, test_spans, tracer
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_request_header_value(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -365,7 +365,7 @@ def test_django_sqli_http_request_header_value(client, test_spans, tracer):
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_iast_disabled_sqli_http_request_header_value(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -385,7 +385,7 @@ def test_django_iast_disabled_sqli_http_request_header_value(client, test_spans,
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_request_header_name(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -423,7 +423,7 @@ def test_django_sqli_http_request_header_name(client, test_spans, tracer):
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_iast_disabled_sqli_http_request_header_name(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -443,7 +443,7 @@ def test_django_iast_disabled_sqli_http_request_header_name(client, test_spans, 
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_path_parameter(client, test_spans, tracer):
     root_span, response = _aux_appsec_get_root_span(
         client,
@@ -478,7 +478,7 @@ def test_django_sqli_http_path_parameter(client, test_spans, tracer):
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_iast_disabled_sqli_http_path_parameter(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -496,7 +496,7 @@ def test_django_iast_disabled_sqli_http_path_parameter(client, test_spans, trace
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_cookies_name(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -536,7 +536,7 @@ def test_django_sqli_http_cookies_name(client, test_spans, tracer):
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_iast_disabled_sqli_http_cookies_name(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -554,7 +554,7 @@ def test_django_iast_disabled_sqli_http_cookies_name(client, test_spans, tracer)
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_cookies_value(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -596,7 +596,7 @@ def test_django_sqli_http_cookies_value(client, test_spans, tracer):
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_iast_disabled_sqli_http_cookies_value(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -621,7 +621,7 @@ def test_django_iast_disabled_sqli_http_cookies_value(client, test_spans, tracer
     ],
 )
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_sqli_http_body(client, test_spans, tracer, payload, content_type):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -678,7 +678,7 @@ def test_django_sqli_http_body(client, test_spans, tracer, payload, content_type
     ],
 )
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_tainted_http_body_empty(client, test_spans, tracer, payload, content_type, deduplication, sampling):
     with override_global_config(
         dict(_iast_enabled=True, _deduplication_enabled=deduplication, _iast_request_sampling=sampling)
@@ -698,7 +698,7 @@ def test_django_tainted_http_body_empty(client, test_spans, tracer, payload, con
 
 
 @pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_iast_disabled_sqli_http_body(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=False)):
         root_span, response = _aux_appsec_get_root_span(
@@ -716,7 +716,7 @@ def test_django_iast_disabled_sqli_http_body(client, test_spans, tracer):
         assert response.content == b"master"
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_querydict(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True)):
         root_span, response = _aux_appsec_get_root_span(
@@ -734,7 +734,7 @@ def test_django_querydict(client, test_spans, tracer):
         )
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_command_injection(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -767,7 +767,7 @@ def test_django_command_injection(client, test_spans, tracer):
         assert loaded["vulnerabilities"][0]["location"]["path"] == TEST_FILE
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_header_injection(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -795,7 +795,7 @@ def test_django_header_injection(client, test_spans, tracer):
         assert loaded["vulnerabilities"][0]["location"]["path"] == TEST_FILE
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_insecure_cookie(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -820,7 +820,7 @@ def test_django_insecure_cookie(client, test_spans, tracer):
         assert vulnerability["hash"]
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_insecure_cookie_secure(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -836,7 +836,7 @@ def test_django_insecure_cookie_secure(client, test_spans, tracer):
         assert root_span.get_tag(IAST.JSON) is None
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_insecure_cookie_empty_cookie(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -852,7 +852,7 @@ def test_django_insecure_cookie_empty_cookie(client, test_spans, tracer):
         assert root_span.get_tag(IAST.JSON) is None
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_insecure_cookie_2_insecure_1_secure(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -870,7 +870,7 @@ def test_django_insecure_cookie_2_insecure_1_secure(client, test_spans, tracer):
         assert len(loaded["vulnerabilities"]) == 2
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_insecure_cookie_special_characters(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _iast_deduplication_enabled=False)):
         oce.reconfigure()
@@ -895,7 +895,7 @@ def test_django_insecure_cookie_special_characters(client, test_spans, tracer):
         assert vulnerability["hash"]
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_stacktrace_leak(client, test_spans, tracer):
     with override_global_config(dict(_iast_enabled=True, _deduplication_enabled=False)):
         oce.reconfigure()
@@ -931,7 +931,7 @@ def debug_mode():
     settings.DEBUG = original_debug
 
 
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
+@pytest.mark.skipif(not asm_config._iast_supported, reason="Python version not supported by IAST")
 def test_django_stacktrace_from_technical_500_response(client, test_spans, tracer, debug_mode):
     with override_global_config(dict(_iast_enabled=True, _deduplication_enabled=False)):
         with tracer.trace("test", span_type=SpanTypes.WEB, service="test") as span:
