@@ -2,9 +2,9 @@ import mock
 import pytest
 
 from ddtrace.appsec._iast import oce
-from ddtrace.appsec._iast._utils import _is_python_version_supported
 from ddtrace.contrib.dbapi import TracedCursor
 from ddtrace.settings import Config
+from ddtrace.settings.asm import config as asm_config
 from ddtrace.settings.integration import IntegrationConfig
 from ddtrace.trace import Pin
 from tests.appsec.iast.conftest import _end_iast_context_and_oce
@@ -33,7 +33,7 @@ class TestTracedCursor(TracerTestCase):
         ):
             _end_iast_context_and_oce()
 
-    @pytest.mark.skipif(not _is_python_version_supported(), reason="IAST compatible versions")
+    @pytest.mark.skipif(not asm_config._iast_supported, reason="IAST compatible versions")
     def test_tainted_query(self):
         from ddtrace.appsec._iast._taint_tracking import OriginType
         from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
@@ -56,7 +56,7 @@ class TestTracedCursor(TracerTestCase):
 
             mock_sql_injection_report.assert_called_once_with(evidence_value=query, dialect="sqlite")
 
-    @pytest.mark.skipif(not _is_python_version_supported(), reason="IAST compatible versions")
+    @pytest.mark.skipif(not asm_config._iast_supported, reason="IAST compatible versions")
     def test_tainted_query_args(self):
         from ddtrace.appsec._iast._taint_tracking import OriginType
         from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
@@ -79,7 +79,7 @@ class TestTracedCursor(TracerTestCase):
 
             mock_sql_injection_report.assert_not_called()
 
-    @pytest.mark.skipif(not _is_python_version_supported(), reason="IAST compatible versions")
+    @pytest.mark.skipif(not asm_config._iast_supported, reason="IAST compatible versions")
     def test_untainted_query(self):
         with mock.patch(
             "ddtrace.appsec._iast.taint_sinks.sql_injection.SqlInjection.report"
@@ -94,7 +94,7 @@ class TestTracedCursor(TracerTestCase):
 
             mock_sql_injection_report.assert_not_called()
 
-    @pytest.mark.skipif(not _is_python_version_supported(), reason="IAST compatible versions")
+    @pytest.mark.skipif(not asm_config._iast_supported, reason="IAST compatible versions")
     def test_untainted_query_and_args(self):
         with mock.patch(
             "ddtrace.appsec._iast.taint_sinks.sql_injection.SqlInjection.report"
@@ -110,7 +110,7 @@ class TestTracedCursor(TracerTestCase):
 
             mock_sql_injection_report.assert_not_called()
 
-    @pytest.mark.skipif(not _is_python_version_supported(), reason="IAST compatible versions")
+    @pytest.mark.skipif(not asm_config._iast_supported, reason="IAST compatible versions")
     def test_tainted_query_iast_disabled(self):
         from ddtrace.appsec._iast._taint_tracking import OriginType
         from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
