@@ -9,8 +9,8 @@ import pytest
 
 @pytest.mark.parametrize("iast_enabled", ["true", "false"])
 @pytest.mark.parametrize("iast_request_sampling", ["100.0", "0.0"])
-@pytest.mark.parametrize("pytest_use_new_plugin", ["true", "false"])
-def test_flask_pytest_iast(iast_enabled, iast_request_sampling, pytest_use_new_plugin):
+@pytest.mark.parametrize("pytest_use_legacy_plugin", ["false", "true"])
+def test_flask_pytest_iast(iast_enabled, iast_request_sampling, pytest_use_legacy_plugin):
     from tests.utils import _build_env
 
     env = _build_env()
@@ -21,7 +21,7 @@ def test_flask_pytest_iast(iast_enabled, iast_request_sampling, pytest_use_new_p
             "DD_TRACE_SQLITE_ENABLED": "0",
             "DD_IAST_ENABLED": iast_enabled,
             "DD_TRACE_DEBUG": "true",
-            "DD_PYTEST_USE_NEW_PLUGIN_BETA": pytest_use_new_plugin,
+            "_DD_PYTEST_USE_LEGACY_PLUGIN": pytest_use_legacy_plugin,
             "DD_IAST_REQUEST_SAMPLING": iast_request_sampling,
             # "DD_API_KEY": "invalidapikey",
             # "DD_CIVISIBILITY_AGENTLESS_ENABLED": "1",
